+++ CHANGES.rst
@@ -67,6 +67,9 @@ Changelog
 - PLIP 13260: Migration contact-info to ``z3c.form`` and make it highly
   customizeable.
   [timitos, saily]
+  
+- PLIP 13260: Migrate author page to browser views/z3c.form (issue #78)
+  [bosim]
 
 
 5.0a1 (2014-03-02)